+++ backend/ingest/ingest_adp.py
@@ -9,21 +9,9 @@ from datetime import datetime, timezone
 import glob
 import pandas as pd
 from backend import config
+from backend.utils import normalize_name
 import logging
 import re
-
-# Setup logging
-logging.basicConfig(level=logging.INFO, format='[%(levelname)s] %(message)s')
-
-def normalize_name(name: str) -> str:
-    """Normalizes player names for consistent matching."""
-    if not isinstance(name, str):
-        return name
-    name = name.lower()
-    name = name.replace("'", "") # Remove apostrophes
-    name = re.sub(r'[^a-z0-9\s]', '', name)  # Remove other non-alphanumeric except spaces
-    name = re.sub(r'(jr|sr|ii|iii|iv)$', '', name)  # Remove common suffixes at end
-    return name.strip()
 
 def _read_adp_parquet(_format: str) -> pd.DataFrame | None:
     """Reads a single ADP parquet file, returning None if not found."""
