+++ backend/ingest/ingest_stats.py
@@ -5,22 +5,10 @@ and saves the data into position-specific Parquet files.
 import nfl_data_py as nfl
 import pandas as pd
 from backend import config
+from backend.utils import normalize_name
 from .stat_columns import player_columns, qb_columns, rb_columns, wr_columns
 import logging
 import re
-
-# Setup logging
-logging.basicConfig(level=logging.INFO, format='[%(levelname)s] %(message)s')
-
-def normalize_name(name: str) -> str:
-    """Normalizes player names for consistent matching."""
-    if not isinstance(name, str):
-        return name
-    name = name.lower()
-    name = name.replace("'", "") # Remove apostrophes
-    name = re.sub(r'[^a-z0-9\s]', '', name)  # Remove other non-alphanumeric except spaces
-    name = re.sub(r'(jr|sr|ii|iii|iv)$', '', name)  # Remove common suffixes at end
-    return name.strip()
 
 def process_position(df: pd.DataFrame, position: str, columns: list[str], sort_by: str, season: int):
     """
